Move `GenreCostProvider` to prod
