+++ Theatrical-Players-Refactoring-Kata/Theatrical-Players-Refactoring-KataTests/StatementPrinterTests.swift
@@ -2,35 +2,6 @@
 import XCTest
 @testable import Theatrical_Players_Refactoring_Kata
 
-struct GenreCostProvider: GenreAmountProvider {
-    func amountFor(genre: String) throws -> AmountCalculator {
-        switch (genre) {
-        case "tragedy" :
-            return { attendance in
-                var result = 40000
-                if (attendance > 30) {
-                    result += 1000 * (attendance - 30)
-                }
-                return result / 100
-            }
-        case "comedy" :
-            return { attendance in
-                var result = 30000
-                if (attendance > 20) {
-                    result += 10000 + 500 * (attendance - 20)
-                }
-                return (result + 300 * attendance) / 100
-            }
-        default:
-            throw GenreError.newGenre
-        }
-    }
-    
-    enum GenreError: Error {
-        case newGenre
-    }
-}
-
 class GenreAmountProviderTests: XCTestCase {
     func test_amountFor_throwsErrorOnNewPlayGenre() throws {
         let sut = GenreCostProvider()
